move axiosWithAuth to utils
